refactor: split nav types, config, and NavSection out of Header.tsx
- Move NavLink / NavSubmenu / NavSectionConfig / NavPagesKey into types.ts
- Move NAV_SECTIONS + EMPTY_PAGES_MAP into components/ui/navSections.ts
- Extract <NavSection> into its own components/ui/NavSection.tsx (with the PageLink helper)
- Header.tsx now only contains the Header component itself
